refactoring (creating helpers.ts)
